Send any configured server hostkey KRL on connect after userauth
Add a krl-00@openssh.com extension that allows a host to provide a key
revocation list and signature to clients. Clients are expected to verify
that the signature is from the same key used to sign the certificate
presented by the host, and may then update a local key revocation list
with the entries.