Revoke trust in certificates that a host asserts are untrustworthy
In the absence of a centralised key revocation mechanism, it's necessary
for environments making use of SSH host certificates to be able to
revoke trust in compromised certificates that would otherwise still be
valid. This patch implements the client side. A new autorevoked_certs
file is introduced, and on connection a server can send a key revocation
list with a signature. If that list is signed with the same key that
signed the certificate presented by the host, any revoked certificate
identifiers included in that KRL corresponding to that CA key will be
merged into autorevoked_certs.